# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/link/": {
            "get": {
                "description": "Returns a list of all available links.",
                "produces": [
                    "application/json"
                ],
                "summary": "Getting all links",
                "responses": {
                    "200": {
                        "description": "List of links",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_link.Link"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing reference by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update link",
                "parameters": [
                    {
                        "description": "Query with link update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_link.Link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated link",
                        "schema": {
                            "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_link.Link"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new link with an optional short URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Link",
                "parameters": [
                    {
                        "description": "Query with Link data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_link.Link"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Link",
                        "schema": {
                            "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_link.Link"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/api/v1/link/{id}/": {
            "get": {
                "description": "Returns information about the link by the short link.",
                "produces": [
                    "application/json"
                ],
                "summary": "Getting a link by short link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short link",
                        "name": "shortLink",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link info",
                        "schema": {
                            "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_link.Link"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an existing reference by ID",
                "summary": "Delete link from db",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/api/v1/status/": {
            "get": {
                "description": "Get the status of the API",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the status",
                "operationId": "Status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_urcop_go-fiber-template_internal_web_status.Response"
                        }
                    }
                }
            }
        },
        "/r/{redirect}": {
            "get": {
                "description": "Redirects to the target URL associated with the given short link.",
                "produces": [
                    "application/json"
                ],
                "summary": "Redirect to a URL",
                "operationId": "redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short link for redirection",
                        "name": "redirect",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to the target URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Error response with 404 status code",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "github.com_urcop_go-fiber-template_internal_web_link.Link": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "clicked": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "random": {
                    "type": "boolean"
                },
                "short_link": {
                    "type": "string"
                }
            }
        },
        "github.com_urcop_go-fiber-template_internal_web_status.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_web_link.Link": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "clicked": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "random": {
                    "type": "boolean"
                },
                "short_link": {
                    "type": "string"
                }
            }
        },
        "internal_web_status.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}